chore: moving files to help overall understanding
